+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - N/A
 
+## [0.0.5] - 2020-09-09
+### Added
+- Add `chassis bootdev pxe|disk|cdrom` command support
+- Add `chassis bootparam get 5` command support
+
 ## [0.0.4] - 2020-09-06
 ### Added
 - Add docker support
@@ -37,7 +42,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Project starts based on the copy of [VirtualBMC 2.1.0.dev](https://github.com/openstack/virtualbmc/commit/c4c8edb66bc49fcb1b8fb41af77546e06d2e8bce)
 
 
-[Unreleased]: https://github.com/kurokobo/virtualbmc-for-vsphere/compare/0.0.4...HEAD
+[Unreleased]: https://github.com/kurokobo/virtualbmc-for-vsphere/compare/0.0.5...HEAD
+[0.0.5]: https://github.com/kurokobo/virtualbmc-for-vsphere/compare/0.0.4...0.0.5
 [0.0.4]: https://github.com/kurokobo/virtualbmc-for-vsphere/compare/0.0.3...0.0.4
 [0.0.3]: https://github.com/kurokobo/virtualbmc-for-vsphere/compare/0.0.2...0.0.3
 [0.0.2]: https://github.com/kurokobo/virtualbmc-for-vsphere/compare/0.0.1...0.0.2